+++ apps/frontend/src/app/modules/scenarios/scenarios.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
 
 import { ScenariosRoutingModule } from './scenarios-routing.module';
 import { ScenariosComponent } from './pages/scenarios/scenarios.component';
@@ -8,6 +9,11 @@ import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   declarations: [ScenariosComponent, ScenarioComponent],
-  imports: [CommonModule, ScenariosRoutingModule, SharedModule],
+  imports: [
+    CommonModule,
+    ScenariosRoutingModule,
+    SharedModule,
+    ReactiveFormsModule,
+  ],
 })
 export class ScenariosModule {}